Refactor loading of ENVs
